requirements.txt
fastapi
pydantic
uvicorn
sqlalchemy
pymysql
pthon-dotenv
python-jose
passlib
fastapi-jwt-auth
cryptography
bcrypt